+++ src/Result.php
@@ -2,19 +2,18 @@
 
 namespace Okipa\LaravelTable;
 
-use Closure;
 use Okipa\LaravelTable\Traits\Result\HasClasses;
+use Okipa\LaravelTable\Traits\Result\HasCustomHtml;
 use Okipa\LaravelTable\Traits\Result\HasTitle;
 
 class Result
 {
     use HasClasses;
     use HasTitle;
+    use HasCustomHtml;
 
     public function __construct()
     {
-        $this->classes = config('laravel-table.classes.results');
+        $this->initializeClasses();
     }
-
-
 }